backend/requirements.txt
asgiref
dj-database-url==0.5.0
django==4.0.1
djangorestframework==3.12.2
gunicorn==20.0.4
joblib==1.0.1
numpy==1.24.2
psycopg2==2.9.5
pytz==2021.1
scikit-learn==1.1.3
scipy==1.10.1
sqlparse
threadpoolctl==2.1.0
whitenoise==5.2.0
djongo==1.3.6
